factor out the json utils
